+++ src/main/java/com/adoptapet/adoptapet/Services/EntityServices/PetService.java
@@ -3,6 +3,8 @@ package com.adoptapet.adoptapet.Services.EntityServices;
 import com.adoptapet.adoptapet.Dtos.PetDto;
 import com.adoptapet.adoptapet.Entities.Pet.Pet;
 import com.adoptapet.adoptapet.Entities.Shelter;
+import com.adoptapet.adoptapet.Exceptions.PetExceptions.PetAlreadyExistsException;
+import com.adoptapet.adoptapet.Exceptions.PetExceptions.PetNotFoundException;
 import com.adoptapet.adoptapet.Mappers.PetMapper;
 import com.adoptapet.adoptapet.Repositories.PetRepository;
 import com.adoptapet.adoptapet.Repositories.ShelterRepository;
@@ -18,10 +20,9 @@ import java.util.List;
 
 @Service
 public class PetService {
+    private static final int PAGE_SIZE = 10;
     private final PetRepository petRepository;
-
     private final ShelterRepository shelterRepository;
-
     private final PetMapper petMapper;
 
     @Autowired
@@ -37,22 +38,29 @@ public class PetService {
     }
 
     public PetDto getPet(int petId) {
-        Pet pet = petRepository.findById(petId).orElse(null);
+        if (petRepository.findById(petId).isEmpty())
+            throw new PetNotFoundException();
+        Pet pet = petRepository.findById(petId).get();
         return petMapper.toDto(pet);
     }
 
     public void addPet(PetDto petDto) {
+        if (petRepository.findById(petDto.getId()).isPresent())
+            throw new PetAlreadyExistsException();
         Pet pet = petMapper.toEntity(petDto);
         petRepository.save(pet);
     }
 
     public void updatePet(PetDto petDto) {
-        Pet pet = petRepository.findById(petDto.getId()).orElse(null);
-        petMapper.partialUpdate(petDto, pet);
+        if (petRepository.findById(petDto.getId()).isEmpty())
+            throw new PetNotFoundException();
+        Pet pet = petRepository.findById(petDto.getId()).get();
         petRepository.save(pet);
     }
 
     public void deletePet(int id) {
+        if (petRepository.findById(id).isEmpty())
+            throw new PetNotFoundException();
         petRepository.deleteById(id);
     }
 
@@ -77,7 +85,7 @@ public class PetService {
     }
 
     public Page<Pet> findAllPets(Specification<Pet> specification, int pageCount) {
-        Pageable pageable = PageRequest.of(pageCount, 10);
+        Pageable pageable = PageRequest.of(pageCount, PAGE_SIZE);
         return petRepository.findAll(specification, pageable);
     }
 }
